requirements.txt
flask==3.0.3
gunicorn==22.0.0
apscheduler==3.10.4
anthropic==0.40.0
fal-client==0.5.0
requests==2.32.3
gspread==6.1.2
google-auth==2.30.0
